LPS-10911 Rollback
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@57422 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portal-impl/src/com/liferay/portal/tools/deploy/BaseDeployer.java
@@ -236,9 +236,12 @@ public class BaseDeployer {
 		throws Exception {
 
 		File file = new File(DeployUtil.getResourcePath(fileName));
+		File targetFile = new File(targetDir + "/" + fileName);
 
-		CopyTask.copyFile(
-			file, new File(targetDir), filterMap, overwrite, true);
+		if (!targetFile.exists()) {
+			CopyTask.copyFile(
+				file, new File(targetDir), filterMap, overwrite, true);
+		}
 	}
 
 	protected void copyJars(File srcFile, PluginPackage pluginPackage)
